+++ src/main.sy
@@ -1,1 +1,2 @@
-if (3 > 2){x = 2} else {x = 3}
+x = 3
+syout(x)
